[indexPattern/flattenSearchResponse] flatten conflicting objects
When flattening hits containing conflicting types, continue to
recursively flatten the object if possible, rather than simply ending
with a complex object.  This fixes a problem when a conflicting type
effectively "hides" nested fields.  For example, if "a" is conflicted
(because sometimes it's an object and sometimes it's a native value),
without this change "a.x" will be hidden from the results, even though
it's only an object-level conflict.  The "a.x" field itself should
appear in the flatten hit as a separate field.