[object.routes] move away from precomputed routes
A circular dependency was discovered in the Doc AppController tests, because it was stubbing
$route and required the IndexPattern class. The dependency path was something like:

$route -> IndexPattern -> Field -> kbnUrl -> $route

While investigating solutions for this I noticed two things: we require kbnUrl both
in the object classes where urls are precomputed and in the controllers/directives
where we redirect/change the url, we needed error-prone logic to recompute the urls
when their dependencies change.

To simplify things and solve the circular dep, the indexPattern and field objects now just
expose a .routes property that has named route patterns. These patterns expect to receive
the object they describe in order to create the route. kbnUrl was also updated with helper
functions for this:

```
kbnUrl.getRouteUrl(indexPattern, 'edit') -> '#/settings/indices/logstash-*'
kbnUrl.redirectToRoute(indexPattern, 'indexedFields')
kbnUrl.changeToRoute(field, 'edit')
```