[fieldFormat] implement highlighting
Asking for a formatted field of contentType 'html' will now highlight the formatted html
when appropriate. In order to do this the format converter function must be passed
the field and hit that corelate to the value. This makes the code read a bit like this:

var convert = field.format.getConverterFor('html');
var formatted = convert(value, field, hit);

The field and hit arguments are not required if highlighting is not desired (like when
formatting results that are not search hits).