[node] downgrade to 0.12.7
While working to reintroduce shrinkwrap into our workflow it became
apparent that we needed to be using npm3 but unfortunately node-gyp
is not compatible with iojs, and the iojs distros actually contain a
patched version. This means that you can only use iojs with the npm
it ships with.

When trying to find a solution for this issue, we contemplated not
using npm3, but that would mean that we can't use npm-shrinkwrap and
decided that since we are using babel anyway that switching back to
node was the best option.

The only code that must be updated to work in the new node 0.12 env
is lodash templates that use the template string syntax inside their
source (Since 0.12 does not support template strings).