Do not flicker "no results" while preparing requests
The recent changes to make segmented request preparation asynchronous
resulted in a regression where the "no results found" screen was being
displayed for a moment before being replaced with the "searching..."
screen. This update ensures that the state where requests are being
prepared is represented by a loading screen as well.