Treat no index in msearch as empty instead of global
Prior to this, if you were to pass an empty index list to the search
fetch strategy in courier, kibana would perform an msearch on all
indexes in elasticsearch. This commit implements the desired behavior
that no defined index list means there are no indexes to query, and thus
is treated like an empty response.