Remove ability to create kibana index via proxy
The kibana index gets created via the server healthcheck during startup,
so the proxy is no longer used to create the index. The route that was
allowing the index to be created (or deleted) is, however, used for
managing contents of that index such as index patterns and advanced
settings, so we can't get rid of it entirely.

Fixes #5792