Backport PR #7490 ---------
**Commit 1:**
Build artifact names for unified release process

These changes are necessary for Kibana to be compatible with Elastic's
unified release process from 5.0 onward. The way artifacts get created
has not changed, but the naming conventions have.

* Original sha: c550bbcc958d74d0e5456d5cebfc81750dd2d54e
* Authored by Court Ewing <court@epixa.com> on 2016-06-16T11:58:31Z

**Commit 2:**
README is 5.0.0-alpha4 to match package.json

We bumped the version in package.json, but we never changed it in the
readme.

* Original sha: 55c499395718ed46f131f9a57927468494b9e35e
* Authored by Court Ewing <court@epixa.com> on 2016-06-16T20:52:58Z