Backport PR #7998 ---------
**Commit 1:**
Don't use deprecated config key `elasticsearch_url`

When following the description from the docs to set the elasticsearch url I got a warning that the key was deprecated. This updates the docs to use the non-deprecated key. I haven't checked if this needs to be updated in other parts of the documentation too.

* Original sha: 3544b66dfa4f9d5523f29e0a69faee1a3a9d6c37
* Authored by Christoph Büscher <christoph@elastic.co> on 2016-08-15T12:14:33Z
* Committed by GitHub <noreply@github.com> on 2016-08-15T12:14:33Z