Only show asc sort terms agg deprecation once per route
The deprecation notification when using terms aggregations that are
using ascending order by count was appearing too frequently, often
multiple times on the same screen just when performing common UI
actions that may not have anything specifically to do with the sort
order of the terms aggregation.

The notification will now appear once whenever the route changes and the
offending terms aggregation is present, but it will not render again
until at least the route changes.